+++ components/compositing/pipeline.rs
@@ -14,7 +14,6 @@ use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
 use ipc_channel::router::ROUTER;
 use layers::geometry::DevicePixel;
 use layout_traits::{LayoutControlChan, LayoutTaskFactory};
-use msg::compositor_msg::ScriptToCompositorMsg;
 use msg::constellation_msg::PaintMsg;
 use msg::constellation_msg::ScriptMsg as ConstellationMsg;
 use msg::constellation_msg::{ConstellationChan, Failure, FrameId, PipelineId, SubpageId};
@@ -27,7 +26,7 @@ use profile_traits::mem as profile_mem;
 use profile_traits::time;
 use script_traits::{ConstellationControlMsg, InitialScriptState};
 use script_traits::{LayoutControlMsg, NewLayoutInfo, ScriptTaskFactory};
-use script_traits::{TimerEventRequest};
+use script_traits::{TimerEventRequest, ScriptToCompositorMsg};
 use std::mem;
 use std::sync::mpsc::{Receiver, Sender, channel};
 use std::thread;
